pyproject.toml
[tool.poetry]
name = "cyberpunk"
version = "0.2.0"
description = "Audio Processing Server"
authors = ["Johannes Naylor <jonaylor89@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/jonaylor89/cyberpunk"
repository = "https://github.com/jonaylor89/cyberpunk"
keywords = ["audio", "server"]

[tool.poetry.dependencies]
python = "^3.6"
pydub = "^0.25.1"
Flask = "^2.0.2"
python-dotenv = "^0.19.2"
boto3 = "^1.20.54"
gunicorn = "^20.1.0"
requests = "^2.27.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.931"
types-requests = "^2.27.9"
pydub-stubs = "^0.25.1"
boto3-stubs = "^1.20.54"
types-PyYAML = "^6.0.4"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
